+++ surfsplit/src/Process.hpp
@@ -15,7 +15,7 @@ std::vector<Rect> CropEmpty( const std::vector<Rect>& rects, Bitmap* bmp );
 std::vector<Rect> MergeHorizontal( const std::vector<Rect>& rects );
 std::vector<Rect> MergeVertical( const std::vector<Rect>& rects );
 
-std::vector<float> CalcBroadDuplicates( const std::vector<Rect>& rects, Bitmap* bmp );
+std::vector<int> CalcBroadDuplicates( const std::vector<Rect>& rects, Bitmap* bmp );
 bool AreExactDuplicates( const Rect& rect1, const Rect& rect2, Bitmap* bmp );
 
 #endif
